+++ src/MPU_Mode.cpp
@@ -12,7 +12,7 @@
  * with the License. You may obtain a copy of the License at
  * http://www.apache.org/licenses/LICENSE-2.0
  */
-#include "MPU_D1_mini.h"
+#include "MiP_Power_Up_-_D1_mini.h"
 
 // Implement the constructor to store the MiP reference.
 MiP_Mode::MiP_Mode(MiP& mip) : m_mip(mip) {}